add generic functions for array test helpers. rename compare to results
